QuadMap based on the OctoMap 1.10.0 implementation

+++ quadmap/include/superray_quadmap/QuadTreeBaseImpl.hxx
@@ -1041,7 +1041,7 @@ namespace quadmap {
 		float step_size = this->resolution * pow(2, tree_depth - depth);
 		for (int i = 0; i < 2; ++i) {
             diff[i] = pmax_clamped(i) - pmin_clamped(i);
-			steps[i] = floor(diff[i] / step_size);
+			steps[i] = static_cast<unsigned int>(floor(diff[i] / step_size));
 			//      std::cout << "bbx " << i << " size: " << diff[i] << " " << steps[i] << " steps\n";
 		}
